feat(week3-task6): split app.tsx into smaller components and add folders for routers and providers
